Create README.md
As an industrial engineer, we took statistics and software courses in our lessons. I decided that these subjects were more suitable for me and I continue to work on it.
Professions that I think I will do with pleasure and long-term in my business life are data analyst, business analyst and software developer.
I continue to develop myself in Python, JAVA and SQL languages.